+++ src/WelcomingSection/WelcomingSection.scss
